refactor: consolidate MCP and dump trees, standardize util imports
- Move all MCP code under `src/adapters/mcp/` (folder with `index.ts` +
  `build-server.ts`, `stringify.ts`, `to-json-schema.ts`, `transports.ts`,
  `__tests__/`), replacing the split between `src/adapters/mcp.ts` and
  `src/node/mcp/`.
- Move dump code under `src/rpc/dump/` (`index.ts`, `error.ts`, `static.ts`,
  `__tests__/`), pulling `src/node/static-dump.ts` in alongside the rest.
  `collectStaticRpcDump` and `StaticRpcDump*` types now ship from
  `devframe/rpc` instead of `devframe/node`.
- Replace relative `../utils/*` imports with `devframe/utils/*` across the
  package; add the convention to AGENTS.md.
- Update build config (`tsdown.config.ts`, `alias.ts`, `tsconfig.base.json`)
  and error-doc source links to match the new paths.
