+++ packages/devframe/src/adapters/cli.ts
@@ -3,7 +3,7 @@ import type { H3 } from 'h3'
 import type { DevframeDefinition } from '../types/devframe'
 import process from 'node:process'
 import cac from 'cac'
-import { colors as c } from '../utils/colors'
+import { colors as c } from 'devframe/utils/colors'
 import { createBuild } from './build'
 import { createDevServer, resolveDevServerPort } from './dev'
 import { flagKeyToOption, isBooleanFlag, parseCliFlags } from './flags'
