+++ packages/devframe/src/adapters/mcp/__tests__/stringify.test.ts
