+++ packages/devframe/src/adapters/mcp/__tests__/to-json-schema.test.ts
