+++ packages/devframe/src/adapters/mcp/stringify.ts
