+++ packages/devframe/src/adapters/mcp/to-json-schema.ts
