+++ packages/devframe/src/node/index.ts
@@ -8,6 +8,5 @@ export * from './host-views'
 export * from './rpc-shared-state'
 export * from './rpc-streaming'
 export * from './server'
-export * from './static-dump'
 export * from './storage'
 export * from './utils'
