+++ packages/devframe/src/rpc/dump/__tests__/dump.test.ts
@@ -1,7 +1,7 @@
-import type { RpcDumpRecord } from './types'
+import type { RpcDumpRecord } from '../../types'
 import * as v from 'valibot'
 import { describe, expect, it } from 'vitest'
-import { createClientFromDump, createDefineWrapperWithContext, defineRpcFunction, dumpFunctions } from '.'
+import { createClientFromDump, createDefineWrapperWithContext, defineRpcFunction, dumpFunctions } from '../..'
 
 describe('dumps', () => {
   it('should collect dumps from definition', async () => {
