+++ packages/devframe/src/rpc/dump/static.ts
