+++ packages/devframe/src/rpc/index.ts
@@ -1,7 +1,9 @@
 export * from './cache'
 export * from './collector'
 export * from './define'
-export * from './dumps'
+export * from './dump'
+export * from './dump/error'
+export * from './dump/static'
 export * from './handler'
 export * from './serialization'
 export * from './types'
