+++ packages/devframe/src/types/rpc.ts
@@ -1,8 +1,8 @@
 import type { BirpcReturn } from 'birpc'
 import type { RpcFunctionsCollectorBase } from 'devframe/rpc'
 import type { DevToolsNodeRpcSessionMeta } from 'devframe/rpc/transports/ws-server'
-import type { SharedState } from '../utils/shared-state'
-import type { StreamReader, StreamSink } from '../utils/streaming-channel'
+import type { SharedState } from 'devframe/utils/shared-state'
+import type { StreamReader, StreamSink } from 'devframe/utils/streaming-channel'
 import type { DevToolsNodeContext } from './context'
 import type { DevToolsRpcClientFunctions, DevToolsRpcServerFunctions, DevToolsRpcSharedStates } from './rpc-augments'
